🧹 back: move all geospatial function in geo_ files
Group geospatial functions together.
